+++ Stripe/Project.swift
@@ -88,6 +88,8 @@ let project = Project(
                 .package(product: "OCMock"),
                 .package(product: "iOSSnapshotTestCase"),
                 .project(target: "StripeCoreTestUtils", path: "//StripeCore"),
+                .project(target: "StripePaymentsObjcTestUtils", path: "//StripePayments"),
+                .project(target: "StripePaymentsTestUtils", path: "//StripePayments"),
                 .project(target: "StripePayments", path: "//StripePayments"),
                 .project(target: "StripePaymentsUI", path: "//StripePaymentsUI"),
                 .project(target: "StripePaymentSheet", path: "//StripePaymentSheet"),