+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentOption+Images.swift
@@ -158,6 +158,10 @@ extension STPPaymentMethodType {
                     return .pm_type_upi
                 case .cashApp:
                     return .pm_type_cashapp
+                case .blik:
+                    return .pm_type_blik
+                case .grabPay:
+                    return .pm_type_grabpay
                 default:
                     return nil
                 }