+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheetFormFactory/PaymentSheetFormFactory.swift
@@ -155,6 +155,10 @@ class PaymentSheetFormFactory {
             additionalElements = [makePaypalMandate()]
         } else if paymentMethod.stpPaymentMethodType == .bancontact {
             return makeBancontact()
+        } else if paymentMethod.stpPaymentMethodType == .blik {
+            return makeDefaultsApplierWrapper(for: makeBLIK())
+        } else if paymentMethod.stpPaymentMethodType == .grabPay {
+            return FormElement(elements: [], theme: theme)
         }
 
         guard let spec = specFromJSONProvider() else {