+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/STPApplePayContext+PaymentSheet.swift
@@ -55,7 +55,7 @@ private class ApplePayContextClosureDelegate: NSObject, ApplePayContextDelegate 
                 return
             }
             let shouldSavePaymentMethod = false // Apple Pay doesn't present the customer the choice to choose to save their payment method
-            intentConfig.confirmHandler(stpPaymentMethod, shouldSavePaymentMethod) { result in
+            intentConfig.confirmHandler(stpPaymentMethod, nil, shouldSavePaymentMethod) { result in
                 switch result {
                 case .success(let clientSecret):
                     guard clientSecret != PaymentSheet.IntentConfiguration.COMPLETE_WITHOUT_CONFIRMING_INTENT else {