+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/ViewControllers/PollingViewModel.swift
@@ -15,11 +15,13 @@ import UIKit
 class PollingViewModel {
 
     let paymentMethodType: STPPaymentMethodType
-    let supportedPaymentMethods: [STPPaymentMethodType] = [.UPI]
+    let supportedPaymentMethods: [STPPaymentMethodType] = [.UPI, .blik]
     lazy var CTA: String = {
         switch paymentMethodType {
         case .UPI:
             return .Localized.open_upi_app
+        case .blik:
+            return .Localized.blik_confirm_payment
         default:
             fatalError("Polling CTA has not been implemented for \(paymentMethodType)")
         }
@@ -28,6 +30,8 @@ class PollingViewModel {
         switch paymentMethodType {
         case .UPI:
             return Date().addingTimeInterval(60 * 5) // 5 minutes
+        case .blik:
+            return Date().addingTimeInterval(60) // 60 seconds
         default:
             fatalError("Polling deadline has not been implemented for \(paymentMethodType)")
         }