+++ StripePayments/StripePaymentsTestUtils/StripePaymentsTestUtils.h
@@ -1,0 +1,16 @@
+//
+//  StripePaymentsTestUtils.h
+//  StripePaymentsTestUtils
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for StripePaymentsTestUtils.
+FOUNDATION_EXPORT double StripePaymentsTestUtilsVersionNumber;
+
+//! Project version string for StripePaymentsTestUtils.
+FOUNDATION_EXPORT const unsigned char StripePaymentsTestUtilsVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <StripePaymentsTestUtils/PublicHeader.h>
+
+